fix registration and move to user management library
