+++ SourceCode/SS.Backend/RegistrationAPI/Program.cs
@@ -1,84 +1,108 @@
-using SS.Backend.Services.AccountCreationService;
+using SS.Backend.Services;
+using Microsoft.AspNetCore.Mvc;
+using System.Threading.Tasks;
+using System.Data;
+using SS.Backend.SharedNamespace;
+using SS.Backend.ReservationManagement;
+using SS.Backend.DataAccess;
+using SS.Backend.Security;
+using Microsoft.IdentityModel.Tokens;
+using Microsoft.Net.Http.Headers;
+using SS.Backend.Services.LoggingService;
+using SS.Backend.UserManagement;
+
+using System.Text;
+
+
+
+
+
 var builder = WebApplication.CreateBuilder(args);
 
+
 builder.Services.AddControllers();
+builder.Services.AddEndpointsApiExplorer();
+builder.Services.AddSwaggerGen();
+
+
+
+
+var baseDirectory = AppContext.BaseDirectory;
+var projectRootDirectory = Path.GetFullPath(Path.Combine(baseDirectory, "../../../../../"));
+var configFilePath = Path.Combine(projectRootDirectory, "Configs", "config.local.txt");
+
+
+//Dao Setup
+builder.Services.AddTransient<ConfigService>(provider =>new ConfigService(configFilePath));
+builder.Services.AddTransient<ISqlDAO, SqlDAO>();
+builder.Services.AddTransient<CustomSqlCommandBuilder>();
+
+
+
+
+
+//security
+
+builder.Services.AddTransient<GenOTP>();
+builder.Services.AddTransient<Hashing>();
+builder.Services.AddTransient<Response>();
+builder.Services.AddTransient<LogEntry>();
+builder.Services.AddTransient<ILogTarget, SqlLogTarget>();
+builder.Services.AddTransient<SS.Backend.Services.LoggingService.ILogger,Logger>();
+builder.Services.AddTransient<Logger>();
+builder.Services.AddTransient<SqlDAO>();
+
+
 builder.Services.AddTransient<IAccountCreation, AccountCreation>();
+
+
+
+builder.Services.AddTransient<SSAuthService>(provider =>
+    new SSAuthService(
+        provider.GetRequiredService<GenOTP>(),
+        provider.GetRequiredService<Hashing>(),
+        provider.GetRequiredService<SqlDAO>(),
+        provider.GetRequiredService<Logger>()
+    )
+);
+
+// Learn more about configuring Swagger/OpenAPI
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
 
 var app = builder.Build();
-
-// Manually handle CORS
-app.Use(async (context, next) =>
+app.Use((context, next) =>
 {
-    // Set the necessary headers for CORS
-    context.Response.Headers.Append("Access-Control-Allow-Origin", "*");
-    context.Response.Headers.Append("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE, OPTIONS");
+    
+    context.Response.Headers.Append("Access-Control-Allow-Origin", "http://localhost:3000");
+    context.Response.Headers.Append("Access-Control-Allow-Methods", "GET, POST, PUT, OPTIONS");
     context.Response.Headers.Append("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept, Authorization");
+    context.Response.Headers.Append("Access-Control-Allow-Credentials", "true");
 
-    // Handle the preflight request
+    
     if (context.Request.Method == "OPTIONS")
     {
-        context.Response.StatusCode = 200;
-        await context.Response.CompleteAsync();
-        return;
+        context.Response.Headers.Append("Access-Control-Max-Age", "86400"); 
+        context.Response.StatusCode = 204; 
+        return Task.CompletedTask;
     }
-    await next();
+
+    return next();
 });
+
 
 if (app.Environment.IsDevelopment())
 {
     app.UseSwagger();
     app.UseSwaggerUI();
-    app.UseDeveloperExceptionPage();
+    app.UseDeveloperExceptionPage();    
 }
 
 app.UseHttpsRedirection();
-app.UseAuthorization();
+
+app.UseMiddleware<AuthorizationMiddleware>();
 
 app.MapControllers();
 
 app.Run();
 
-
-
-
-// using SS.Backend.Services.AccountCreationService;
-
-// var builder = WebApplication.CreateBuilder(args);
-
-// builder.Services.AddControllers();
-// builder.Services.AddTransient<IAccountCreation, AccountCreation>();
-
-// // Define a CORS policy
-// const string MyAllowSpecificOrigins = "_myAllowSpecificOrigins";
-
-// builder.Services.AddEndpointsApiExplorer();
-// builder.Services.AddSwaggerGen();
-
-// // Add CORS services
-// builder.Services.AddCors();
-
-// var app = builder.Build();
-
-// // Enable CORS
-// app.UseCors(policy =>
-// {
-//     policy.AllowAnyOrigin();
-//     policy.AllowAnyHeader();
-//     policy.AllowAnyMethod();
-// });
-
-// if (app.Environment.IsDevelopment())
-// {
-//     app.UseSwagger();
-//     app.UseSwaggerUI();
-//     app.UseDeveloperExceptionPage();
-// }
-
-// app.UseHttpsRedirection();
-// app.UseAuthorization();
-
-// app.MapControllers();
-
-// app.Run();
