+++ SourceCode/SS.Backend/RegistrationAPI/Registration.csproj
@@ -8,16 +8,16 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.FileProviders.Physical" Version="8.0.0" />
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.4.0" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\SS.Backend\SS.Backend.Services\SS.Backend.Services.csproj" />
-    <!-- <ProjectReference Include="..\..\SS.Backend\SS.Backend.UserManagement/SS.Backend.UserManagement.csproj" /> -->
-    <!-- <ProjectReference Include="..\..\SS.Backend\SS.Backend.DataAccess\SS.Backend.DataAccess.csproj" /> -->
-    <!-- <ProjectReference Include="..\..\SS.Backend\SS.Backend.Security\SS.Backend.Security.csproj" /> -->
     <ProjectReference Include="..\..\SS.Backend\SS.Backend.SharedNamespace\SS.Backend.SharedNamespace.csproj" />
+    <!-- <ProjectReference Include="..\..\SS.Backend\SS.Backend.Services\SS.Backend.Services.csproj" /> -->
+    <ProjectReference Include="..\..\SS.Backend\SS.Backend.DataAccess\SS.Backend.DataAccess.csproj"/>
+		<ProjectReference Include="..\SS.Backend.Security.AuthN\SS.Backend.Security.AuthN.csproj" />
+		<ProjectReference Include="..\SS.Backend.Security\SS.Backend.Security.csproj" />
+    <ProjectReference Include="..\SS.Backend.UserManagement\SS.Backend.UserManagement.csproj" />
   </ItemGroup>
 
 </Project>
