+++ SourceCode/SS.Backend/SS.Backend.Services/AccountCreationService/IAccountCreation.cs
@@ -1,12 +1,12 @@
-using SS.Backend.SharedNamespace;
+// using SS.Backend.SharedNamespace;
 
-namespace SS.Backend.Services.AccountCreationService
-{
-    public interface IAccountCreation
-    {
-        public Task<Response> CreateUserAccount(UserInfo userInfo, CompanyInfo? companyInfo);
-        public Task<Response> ReadUserTable(string tableName);
-        public Task<Response> VerifyAccount(string username);
-        public Task<Response> getEmployeeCompanyID(UserInfo userInfo, string manager_hashedUsername);
-    }
-}
+// namespace SS.Backend.UserManagement
+// {
+//     public interface IAccountCreation
+//     {
+//         public Task<Response> CreateUserAccount(UserInfo userInfo, CompanyInfo? companyInfo);
+//         public Task<Response> ReadUserTable(string tableName);
+//         public Task<Response> VerifyAccount(string username);
+//         public Task<Response> getEmployeeCompanyID(UserInfo userInfo, string manager_hashedUsername);
+//     }
+// }
