+++ SourceCode/SS.Backend/SS.Backend.Tests.AccountCreation/SS.Backend.Tests.AccountCreation.csproj
@@ -14,6 +14,7 @@
     <ProjectReference Include="..\SS.Backend.DataAccess\SS.Backend.DataAccess.csproj" />
     <ProjectReference Include="..\SS.Backend.Services\SS.Backend.Services.csproj" />
     <ProjectReference Include="..\SS.Backend.SharedNamespace\SS.Backend.SharedNamespace.csproj" />
+    <ProjectReference Include="..\SS.Backend.UserManagement\SS.Backend.UserManagement.csproj" />
   </ItemGroup>
