+++ SourceCode/SS.Backend/SS.Backend.UserManagement/Contracts/IUserManagementDao.cs
@@ -11,7 +11,6 @@ namespace SS.Backend.UserManagement
         public Task<Response> createAccountRecoveryRequest(UserRequestModel userRequest, string tableName);
         public  Task<Response> sendRequest(string name, string position);
         public  Task<Response> readTableWhere(string whereClause, object whereClauseval, string tableName);
-        
-
+        public Task<Response> CreateAccount(UserInfo userInfo, CompanyInfo? companyInfo);
     }
 }
