+++ SourceCode/SS.Backend/SS.Backend.UserManagement/SS.Backend.UserManagement.csproj
@@ -10,9 +10,10 @@
 
 
   <ItemGroup>
-  <PackageReference Include="Microsoft.Data.SqlClient" Version="5.1.4" />
+  <PackageReference Include="Microsoft.Data.SqlClient" Version="5.2.0" />
     <ProjectReference Include="..\SS.Backend.DataAccess\SS.Backend.DataAccess.csproj" />
     <ProjectReference Include="..\SS.Backend.SharedNamespace\SS.Backend.SharedNamespace.csproj" />
+    <ProjectReference Include="..\SS.Backend.Services\SS.Backend.Services.csproj" />
   </ItemGroup>
 
 </Project>
