+++ SourceCode/SS.Frontend/Login/index.html
@@ -317,10 +317,6 @@
                                 <option value="Low">Low</option>
                             </select>
                         </div>
-                        <!-- <div>
-                            <label for="taskNotificationSetting">Notification Setting:</label>
-                            <input type="number" id="taskNotificationSetting" name="notificationSetting" required>
-                        </div> -->
                         <button type="submit" id="createTaskBtn" data-testid="submit-task-button">Create Task</button>
                     </form>
